Moved all auxiliary methods to helpers
